move private functions to separate module
