+++ cartoframes/cartoframes.py
@@ -1,125 +1,36 @@
 """
-Monkey patching pandas to add utilities for CARTO tables and maps
+A pandas dataframe interface for working with CARTO maps and tables
+
 Andy Eschbacher and Stuart Lynn, 2017
 
 Project goals
     * Interact with a CARTO table fully within a Jupyter notebook/pandas
       workflow (read and sync dataframe changes)
     * Like geopandas, have a .carto_map() method which gives back the data
-      as a map using carto's maps api and turbocartocss on an optional
+      as a map using carto's maps api and TurboCartoCSS on an optional
       attribute
     * Add CARTO services like the Data Observatory as methods to a dataframe
 
-Features to add:
-    * create a dataframe from scratch
-        * establish cartodb_id
-        * set metadata manually
-        * register with carto
-
-Notes on propagating pandas metadata:
-    * https://github.com/pandas-dev/pandas/issues/2485
-    * geopandas does a good job of propagating metadata, seems to be by
-      subclassing Dataframes:
-      https://github.com/geopandas/geopandas/blob/v0.2.1/geopandas/geodataframe.py#L54
-      similar to what we tried in cartopandas.py.
-      A geodataframe stores it's own metadata:
-      https://github.com/geopandas/geopandas/blob/v0.2.1/geopandas/geodataframe.py#L47
+Features to add: see issues in the repository https://github.com/CartoDB/cartoframes/issues?q=is%3Aopen+is%3Aissue+label%3Aenhancement
 """
 
 # TODO: hook into pandas.core?
 import pandas as pd
+import cartoframes.utils as carto_utils
+import carto
+# TODO: setup carto authentication through a method
 
-# NOTE: `add_meta` not currently used
-def add_meta(self, **kwargs):
-    """
-        Set metadata for a dataframe if none has been already set
-    """
-    for key in kwargs:
-        self._metadata[0][key] = kwargs[key]
-
-def map_dtypes(pgtype):
-    """
-        Map PostgreSQL data types (key) to NumPy/pandas dtypes (value)
-    """
-    # may not be a complete list, could not find SQL API documentation
-    # about data types
-    dtypes = {'number': 'float64',
-              'date': 'datetime64',
-              'string': 'object',
-              'geometry': 'object',
-              'boolean': 'bool'}
-    try:
-        return dtypes[pgtype]
-    except KeyError:
-        return 'object'
-
-def dtype_to_pgtype(dtype, colname):
-    """
-    Map dataframe types to carto postgres types
-    """
-    if colname in ('the_geom', 'the_geom_webmercator'):
-        return 'geometry'
-    else:
-        mapping = {'float64': 'numeric',
-                   'datetime64': 'date',
-                   'object': 'text',
-                   'bool': 'boolean'}
-        try:
-            return mapping[dtype]
-        except KeyError:
-            return 'string'
-
-def transform_schema(pgschema):
-    """
-        Transform schema returned via SQL API to dict for pandas
-    """
-    datatypes = {}
-    for field in pgschema:
-        if 'cartodb_id' in field:
-            continue
-        datatypes[field] = map_dtypes(pgschema[field]['type'])
-    return datatypes
-
-def get_username(baseurl):
-    """
-    NOTE: Not compatible with onprem, etc.
-    """
-    import re
-    m = re.search('https://(.*?).carto.com/api/', baseurl)
-    return m.group(1)
-
-def get_geom_type(sql_auth_client, tablename):
-    """
-        Get the geometry type in tablename for storing in
-        dataframe metadata
-    """
-    geomtypes = {'ST_Point': 'point',
-                 'ST_MultiPoint': 'point',
-                 'ST_LineString': 'line',
-                 'ST_MultiLineString': 'line',
-                 'ST_Polygon': 'polygon',
-                 'ST_MultiPolygon': 'polygon'}
-
-    # NOTE: assumes one geometry type per table
-    result = sql_auth_client.send('''
-        SELECT ST_GeometryType(the_geom) As geomtype
-        FROM "{tablename}"
-        LIMIT 1'''.format(tablename=tablename))
-    try:
-        return geomtypes[result['rows'][0]['geomtype']]
-    except KeyError:
-        print("Warning: cannot create a map from `{tablename}` because this "
-              "table does not have geometries").format(tablename=tablename)
-        return None
 
 # NOTE: this is compatible with v1.0.0 of carto-python client
-# TODO: remove username as a param would be nice.. accessible to write to
-#       metadata from carto python client?
 def read_carto(cdb_client, tablename=None,
                query=None, include_geom=True,
                limit=None, index='cartodb_id', debug=False):
     """Import a table from carto into a pandas dataframe, storing
-       table information in pandas metadata"""
+       table information in pandas metadata.
+       Inputs:
+       :param cdb_client: object CARTO Python SDK authentication client
+       :param
+       """
     from carto.sql import SQLClient
     import json
     sql = SQLClient(cdb_client)
@@ -127,10 +38,11 @@ def read_carto(cdb_client, tablename=None,
     # construct query
     if tablename:
         query = 'SELECT * FROM "{tablename}"'.format(tablename=tablename)
-        geomtype = get_geom_type(sql, tablename)
+        geomtype = carto_utils.get_geom_type(sql, tablename)
         # Add limit if requested
         if limit:
             # NOTE: what if limit is `all` or `none`?
+            # TODO: ensure that this does not cause sync problems
             if (limit >= 0) and isinstance(limit, int):
                 query += ' LIMIT {limit}'.format(limit=limit)
             else:
@@ -138,9 +50,10 @@ def read_carto(cdb_client, tablename=None,
     elif query:
         # NOTE: note yet implemented
         # query = custom_query
-        pass
+        raise NotImplementedError("Creating a cartoframe from a query is not "
+                                  "yet implemented.")
     else:
-        raise NameError("`tablename` or `query` needs to be specified")
+        raise NameError("Either `tablename` or `query` needs to be specified")
 
     if debug:
         print(query)
@@ -152,15 +65,15 @@ def read_carto(cdb_client, tablename=None,
         print(query)
     # TODO: how to handle NaNs deterministically?
     resp = sql.send(query)
-    schema = transform_schema(resp['fields'])
-    _df = pd.DataFrame(resp['rows']).set_index(index).astype(schema)
+    schema = carto_utils.transform_schema(resp['fields'])
+    carto_utils.df = pd.DataFrame(resp['rows']).set_index(index).astype(schema)
 
     # TODO: add table schema to the metadata
     # NOTE: pylint complains that we're accessing a 'protected member
-    #       _metadata of a client class' (appending to _metadata only works
+    #       carto_utils.metadata of a client class' (appending to _metadata only works
     #       with strings, not JSON, so we're serializing here)
     _df._metadata.append(json.dumps({'carto_table': tablename,
-                                     'carto_username': get_username(cdb_client.base_url),
+                                     'carto_username': carto_utils.get_username(cdb_client.base_url),
                                      'carto_include_geom': include_geom,
                                      'carto_limit': limit,
                                      'carto_schema': str(schema),
@@ -168,131 +81,6 @@ def read_carto(cdb_client, tablename=None,
     _df.carto_last_state = _df.copy(deep=True)
     _df.carto_sql_client = sql
     return _df
-
-pd.read_carto = read_carto
-
-
-# TODO: add into sync_carto function as subfunction?
-def process_item(item):
-    """
-      Map NumPy values to PostgreSQL values
-    """
-    from math import isnan
-    if isinstance(item, str):
-        return '\'{}\''.format(item)
-    elif isinstance(item, float):
-        if isnan(item):
-            return 'null'
-        return str(item)
-    return str(item)
-
-def datatype_map(dtype):
-    """
-       map NumPy types to PostgreSQL types
-    """
-    # TODO: add datetype conversion
-    if 'float' in dtype:
-        return 'numeric'
-    elif 'int' in dtype:
-        return 'int'
-    elif 'bool' in dtype:
-        return 'boolean'
-    else:
-        return 'text'
-
-
-def upsert_table(self, df_diff, debug, n_batch=20):
-    import json
-
-    n_items = len(df_diff)
-    queries = []
-
-    for row_num, row in enumerate(df_diff.iteritems()):
-        if debug: print(row)
-        cartodb_id = row[0][0]
-        colname = row[0][1]
-        upsert_query = '''
-        INSERT INTO "{tablename}"("cartodb_id", "{colname}")
-             VALUES ({cartodb_id}, {colval})
-        ON CONFLICT ("cartodb_id")
-        DO UPDATE SET "{colname}" = {colval}
-        WHERE EXCLUDED."cartodb_id" = {cartodb_id};
-        '''.format(tablename=json.loads(self._metadata[0])['carto_table'],
-                   colname=colname,
-                   colval=process_item(self.loc[cartodb_id][colname]),
-                   cartodb_id=cartodb_id).strip().replace('\n', ' ')
-        queries.append(upsert_query)
-        # if debug: print(upsert_query)
-
-        # run batch if at n_batch queries, or at last item
-        if (len(queries) == n_batch) or (row_num == n_items - 1):
-            batchquery = '\n'.join(queries)
-            if debug:
-                print("Num characters in batch query: "
-                      "{}".format(len(batchquery)))
-            resp = self.carto_sql_client.send(batchquery)
-            if debug: print(resp)
-            queries = []
-    return None
-
-
-def drop_col(self, colname, debug):
-    """
-        Drop specified column
-    """
-    import json
-    alter_query = '''
-        ALTER TABLE "{tablename}"
-        DROP COLUMN "{colname}"
-    '''.format(tablename=json.loads(self._metadata[0])['carto_table'],
-               colname=colname)
-
-    if debug: print(alter_query)
-    resp = self.carto_sql_client.send(alter_query)
-    if debug: print(resp)
-    return None
-
-
-def add_col(self, colname, debug):
-    """
-        Alter table add col
-    """
-    import json
-    if debug: print("Create new column {col}".format(col=colname))
-    alter_query = '''
-        ALTER TABLE "{tablename}"
-        ADD COLUMN "{colname}" {datatype}
-    '''.format(tablename=json.loads(self._metadata[0])['carto_table'],
-               colname=colname,
-               datatype=datatype_map(str(self.dtypes[colname])))
-    if debug: print(alter_query)
-    # add column
-    resp = self.carto_sql_client.send(alter_query)
-    if debug: print(resp)
-    # update all the values in that column
-    # NOTE: fails if colval is 'inf' or some other Python or NumPy type
-    for item in self[colname].iteritems():
-        if debug: print(item)
-        update_query = '''
-            UPDATE "{tablename}"
-            SET "{colname}" = {colval}
-            WHERE "cartodb_id" = {cartodb_id};
-        '''.format(tablename=json.loads(self._metadata[0])['carto_table'],
-                   colname=colname,
-                   colval=process_item(item[1]),
-                   cartodb_id=item[0])
-        if debug: print(update_query)
-        resp = self.carto_sql_client.send(update_query)
-        # if debug: print(resp.text)
-    return None
-
-def create_carto_table(self, auth_client, tablename, debug=False):
-    """
-
-    """
-    schema = dict([(col, dtype_to_pgtype(str(dtype), colname))
-                   for col, dtype in zip(self.columns, self.dtypes)])
-    return None
 
 
 # TODO: make less buggy about the diff between NaNs and nulls
@@ -325,14 +113,14 @@ def sync_carto(self, createtable=False, auth_client=None,
     if len(set(self.columns) - set(self.carto_last_state.columns)) > 0:
         newcols = set(self.columns) - set(self.carto_last_state.columns)
         for col in newcols:
-            add_col(self, col, debug)
+            carto_utils.add_col(self, col, debug)
 
     # drop column if needed
     # TODO: extract to function
     if len(set(self.carto_last_state.columns) - set(self.columns)) > 0:
         discardedcols = set(self.carto_last_state.columns) - set(self.columns)
         for col in discardedcols:
-            drop_col(self, col, debug)
+            carto_utils.drop_col(self, col, debug)
 
     # sync updated values
     # TODO: what happens if rows are removed?
@@ -343,113 +131,23 @@ def sync_carto(self, createtable=False, auth_client=None,
         df_diff = (self[common_cols] !=
                    self.carto_last_state[common_cols]).stack()
         df_diff = df_diff[df_diff]
-        upsert_table(self, df_diff, debug)
+        carto_utils.upsert_table(self, df_diff, debug)
 
     # update state of dataframe
     self.carto_last_state = self.copy(deep=True)
     print("Sync completed successfully")
 
-pd.DataFrame.sync_carto = sync_carto
-
-def cartocss_by_geom(geomtype):
-    if geomtype == 'point':
-        markercss = '''
-            #layer {
-              marker-width: 7;
-              marker-fill: %(filltype)s;
-              marker-fill-opacity: 1;
-              marker-allow-overlap: true;
-              marker-line-width: 1;
-              marker-line-color: #FFF;
-              marker-line-opacity: 1;
-            }
-        '''.replace('\n', '')
-        return markercss
-    elif geomtype == 'line':
-        linecss = '''
-            #layer {
-              line-width: 1.5;
-              line-color: %(filltype)s;
-            }
-        '''.replace('\n', '')
-        return linecss
-    elif geomtype == 'polygon':
-        polygoncss = '''
-            #layer {
-              polygon-fill: %(filltype)s;
-              line-width: 0.5;
-              line-color: #FFF;
-              line-opacity: 0.5;
-            }
-        '''.replace('\n', '')
-        return polygoncss
-    return None
-
-
-def get_fillstyle(params):
-    """
-
-    """
-    if 'colorramp' not in params:
-        pass
-    if params['stylecol']:
-        if params['datatype'] == 'float64':
-            fillstyle = ('ramp([{stylecol}], cartocolor(RedOr), '
-                         'quantiles())'.format(stylecol=params['stylecol']))
-        else:
-            fillstyle = ('ramp([{stylecol}], cartocolor(Bold), '
-                         'category(10))'.format(stylecol=params['stylecol']))
-    else:
-        fillstyle = '#f00'
-
-    return fillstyle
-
-
-def get_mapconfig(params):
-    """
-        Anonymous Maps API template for carto.js
-        mapconfig_params = {'username': df_meta['carto_username'],
-                            'tablename': df_meta['carto_table'],
-                            'geomtype': df_meta['geomtype'],
-                            'stylecol': stylecol,
-                            'datatype': str(self[stylecol].dtype)}
-        dtypes one of
-          * quantitative: float64 (float32, int32, int64)
-          * categorical: bool, object
-            * cartocss rule: ramp([room_type], cartocolor(Bold), category(4))
-              dtypes = {'number': 'float64',
-                        'date': 'datetime64',
-                        'string': 'object',
-                        'geometry': 'object',
-                        'boolean': 'bool'}
-        color palettes: https://github.com/CartoDB/CartoColor/blob/master/cartocolor.js
-    """
-
-    cartocss = cartocss_by_geom(params['geomtype']) % {'filltype': get_fillstyle(params)}
-
-    hyperparams = dict({'cartocss': cartocss}, **params)
-    # print(hyperparams)
-
-    mapconfig = '''{"user_name": "%(username)s",
-                    "type": "cartodb",
-                    "sublayers": [{
-                      "type": "http",
-                      "urlTemplate": "http://{s}.basemaps.cartocdn.com/dark_nolabels/{z}/{x}/{y}.png"
-                      }, {
-                      "sql": "select * from %(tablename)s",
-                      "cartocss": "%(cartocss)s"
-                      }],
-                      "subdomains": [ "a", "b", "c" ]
-                      }''' % hyperparams
-
-    return mapconfig
 
 def carto_map(self, interactive=True, stylecol=None):
     """
         Produce and return CARTO maps or iframe embeds
     """
-    # TODO: make this agnostic (for urllib) for python 2 v 3
-    import urllib
+    try:
+        # if Python 3
+        import urllib.parse as urllib
+    except ImportError:
+        # if Python 2
+        import urllib
     import json
     import IPython
 
@@ -472,7 +170,8 @@ def carto_map(self, interactive=True, stylecol=None):
                                      if stylecol in self.columns
                                      else None)}
 
-    mapconfig_params['q'] = urllib.quote(get_mapconfig(mapconfig_params))
+    mapconfig_params['q'] = urllib.quote(
+        carto_utils.get_mapconfig(mapconfig_params))
 
     # print(params)
     url = '?'.join(['/files/cartoframes.html',
@@ -480,4 +179,9 @@ def carto_map(self, interactive=True, stylecol=None):
     iframe = '<iframe src="{url}" width=700 height=350></iframe>'.format(url=url)
     return IPython.display.HTML(iframe)
 
+
+# Monkey patch these methods to pandas
+
+pd.read_carto = read_carto
 pd.DataFrame.carto_map = carto_map
+pd.DataFrame.sync_carto = sync_carto
